Bugfix/ID-30-KeyboardHandler-class (#63)
* multiple controlled car change bug fixed

* Fixed wheel turning bug

* Update: Brake and throttle inputs cancel out each other.

---------

+++ src/AutomatedCar/SystemComponents/Packets/IReadOnlyKeyboardHandlerPacket.cs
@@ -14,7 +14,7 @@
 
         int ThrottlePercentage { get; }
 
-        int WheelPercentage { get; }
+        double WheelPercentage { get; }
 
         SequentialShiftingDirections ShiftUpOrDown { get; }
     }
